created global variable for filter_type
filter_type is used throughout the document. Instead of having to update the variable in each cells, I created a new cell at the top of the document that must be run in order for any of the other cells to functions. This was the filter_type is always up-to-date.